refactor: separate project from project view
