+++ src/data/mod.rs
@@ -2,8 +2,10 @@ mod document;
 mod document_stats;
 mod folder;
 mod markdown_buffer;
+mod project;
 
 pub use document::Document;
 pub use document_stats::DocumentStats;
 pub use folder::Folder;
 pub use markdown_buffer::MarkdownBuffer;
+pub use project::Project;
